feat: Allow compilation without CGO
