+++ scanner/metadata/taglib/taglib_parser.go
@@ -1,3 +1,5 @@
+//+build cgo
+
 package taglib
 
 /*
